client: ban command with real promises...
Still a work in progress but it allows you to know if user was really
banned or if your command failed. We still have to do some tests and
adjust the promise timeout. This commit might be reverted if it is
causing too many issues.

client.ban("channel", "target").then(function(channel, username) {
console.log("SUCCESS");
}, function(err) {
console.log("FAILED " + err + " :-(");
});