client: added auto host bool to hosted event
Added a bool parameter to the hosted event showing if the host was an auto-host. The message I based my changes off of was: ``DBKynd is now auto hosting you.`` 

I'm making an assumption here that the message for auto host with viewers would be ``DBKynd is now auto hosting you for 10 viewers.`` as I don't have an easy method for testing without setting up alt accounts to 'view' and then host.