client: twitchnotify is being deprecated, subscriptions now handles as a USERNOTICE
According to
https://discuss.dev.twitch.tv/t/subscriptions-beta-changes/10023
New subscriptions are now handled as a USERNOTICE similarly to resubs,
and the method object now includes a plan and planName.

Additionally, users can now provide a message upon subscribing, so a msg
and userstate is now provided. This is a lot of parameters, but it
should be backwards compatible with any code that was designed for the
old twitchnotify version.