Fix moderator tracking for USERSTATE messages
The USERSTATE message contains the channel that it's
being sent for but this was not being used. Instead it used
the `this.lastJoined` variable. However, this internal variable
is only updated after the tracking for moderators is completed.

This results in moderator status being assigned to the wrong 
channel and will result in `client.isMod` returning `false` when 
called in an `onJoin` handler (or thereafter).