various: add opts.connection.fetchAgent support
Follow up to 643b2c9 to include an agent option for fetch. I'm not sure
if opts.connection is the right place to put this but if someone is using
agent for the websocket, they may want to use it for HTTP requests tmi.js
may make, so the option should be adjacent in the options.