various: remove api and node-fetch dependency
This is not a backwards compatible change but I don't believe this is a
feature that anyone is actually using. It's not that convenient and a
sore on the library. I'm very happy to completely remove it as well as
a dependency. An error will be thrown if it's called.